Deduplicate fake file
